Add the score background logic. Add the JACKSON support for RestController. Update the in-memory HSQL statements.

+++ src/main/resources/spring2048/sql/schema.sql
@@ -1,12 +1,12 @@
 create table user(
-  user_id integer identity primary key,
+  user_id bigint identity primary key,
   username varchar(20) not null,
   password varchar(20) not null
 );
 
 create table score(
-  score_id integer not null primary key,
-  scoring integer not null,
+  score_id bigint not null primary key,
+  scoring bigint not null,
   score_date datetime not null,
   username varchar(20) not null
 );
